# backend/static/openapi.yaml
openapi: 3.0.1
info:
  title: Ask to PDF
  description: 一个允许用户使用文心一言加载和查询 PDF 的插件。
  version: 1.0.0
servers:
  - url: http://127.0.0.1:3333
paths:
  /pdf/load:
    post:
      operationId: loadPdf
      summary: 解析用户的url
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/loadPdf"
      responses:
        "200":
          description: 加载文件成功！
          content:
            text/event-stream:
              schema:
                  $ref: "#/components/schemas/messageResponse"
  /pdf/query:
    post:
      operationId: queryPdf
      summary: 提取问题
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/queryPdfRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/queryPdfResponse"
      x-history:
        description: 标记在requestBody中mediaType对应的content中用name所标记的字段接收一言交互历史的字段，其内容格式是[{"role":"","content":""},{"role":"","content":""},...] 
        mediaType: application/json
        name: "yiyan_info"
        type: string

components:
  schemas:
    loadPdf:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: 远程文件url地址，用<url><\url>包含
    queryPdfRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: 问题
          # required: true
        yiyan_info:
          type: string 
          description: 历史信息
    queryPdfResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: string
          description: PDF 文档中与用户查询相关的文本块列表。
    messageResponse:
      type: object
      required: [message]
      properties:
        message:
            type: string
            description: 回复信息
  #/documentation/{service}:
  #   get:
  #     operationId: getDocumentation
  #     summary: Get the PDF URLs of a Huawei Cloud service documentation
  #     parameters:
  #     - in: path
  #       name: service
  #       schema:
  #           type: string
  #       required: true
  #       description: The name of the Huawei Cloud service.
  #     responses:
  #       "200":
  #         description: OK
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 urls:
  #                   type: array
  #                   items:
  #                     type: string
  #                   description: The list of PDF URLs.